Рефакторинг функции получения последнего значения для пользователя: перемещение в модуль утилит и удаление дублирующего кода из API и веб-роутеров.
